Fix signed/unsigned comparison in `evt_tag_mem()` and add unit tests for it

+++ lib/eventlog/src/evttags.c
@@ -86,7 +86,7 @@ evt_tag_mem(const char *tag, const void *value, size_t len)
   char *buf = malloc(len + 1);
 
   memcpy(buf, value, len);
-  for (int i = 0; i < len; i++)
+  for (size_t i = 0; i < len; i++)
     {
       if (buf[i] == 0)
         buf[i] = '.';
